+++ src/libraries/System.CodeDom/tests/System.CodeDom.Tests.csproj
@@ -110,7 +110,6 @@
              Link="Common\System\IO\TempFile.cs" />
     <Compile Include="$(CommonTestPath)System\Runtime\Serialization\Formatters\BinaryFormatterHelpers.cs"
              Link="Common\System\Runtime\Serialization\Formatters\BinaryFormatterHelpers.cs" />
-    <Compile Include="System\CodeDom\LineEndingsHelper.cs" />
   </ItemGroup>
   <ItemGroup Condition="'$(TargetFramework)' == '$(NetCoreAppCurrent)'">
     <ProjectReference Include="..\src\System.CodeDom.csproj" />
